pyproject.toml
[tool.poetry]
name = "hydra"
version = "0.1.0"
description = ""
authors = ["Dyllan To <me@dyllan.to>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.20.0"
matplotlib = "^3.3.4"
mplfinance = "^0.12.7-alpha.5"
pyarrow = "^3.0.0"
python-dateutil = "^2.8.1"
tqdm = "^4.56.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hydra = "hydra:main"
